requirements.txt
anyio==3.6.2
bcrypt==4.0.1
click==8.1.3
dnspython==2.3.0
ecdsa==0.18.0
fastapi==0.89.1
h11==0.14.0
httptools==0.5.0
idna==3.4
Jinja2==3.1.2
kafka-python==2.0.2
MarkupSafe==2.1.2
mongoengine==0.26.0
passlib==1.7.4
py-functions==1.0
pyasn1==0.4.8
pydantic==1.10.4
pymongo==4.3.3
python-dotenv==0.21.1
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
typing-extensions==4.4.0
uvicorn==0.20.0



#pip install requirements.txt - it install all packages in the requirements folder